refactor(dashboard): move processing methods into their own package
All processing methods were in the tasks.py file which is supposed to be for celery background tasks such as auto processing of files.
The processing methods were moved into the processing folder respectively.
