+++ AgrifoodIT/celery.py
@@ -5,7 +5,7 @@ from celery import Celery
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'AgrifoodIT.settings')
 
-app = Celery('agrifoodit')
+app = Celery('AgrifoodIT')
 
 # Using a string here means the worker doesn't have to serialize
 # the configuration object to child processes.
